algoview-web/requirements.txt
flask==2.0.1
flask-socketio==5.1.1
plotly==5.1.0
pandas==1.3.3
numpy==1.19.5
python-dotenv==0.19.0
celery==5.1.2
redis==3.5.3 
openpyxl==3.0.9  # для работы с Excel файлами 